+++ rs/p2p/consensus_manager/src/sender.rs
@@ -179,7 +179,6 @@ impl<
         let id = new_artifact.artifact.id();
         let wire_artifact = self.assembler.disassemble_message(new_artifact.artifact);
         let wire_artifact_id = wire_artifact.id();
-        let wire_artifact_attribute = wire_artifact.attribute();
         let entry = self.active_adverts.entry(id.clone());
 
         if let Entry::Vacant(entry) = entry {
@@ -200,7 +199,6 @@ impl<
                     is_latency_sensitive: new_artifact.is_latency_sensitive,
                 },
                 wire_artifact_id,
-                wire_artifact_attribute,
                 child_token_clone,
             );
 
@@ -224,7 +222,6 @@ impl<
             is_latency_sensitive,
         }: ArtifactWithOpt<WireArtifact>,
         id: WireArtifact::Id,
-        attribute: WireArtifact::Attribute,
         cancellation_token: CancellationToken,
     ) {
         let pb_slot_update = pb::SlotUpdate {
@@ -237,10 +234,7 @@ impl<
                 {
                     pb::slot_update::Update::Artifact(pb_artifact.encode_to_vec())
                 } else {
-                    pb::slot_update::Update::Advert(pb::Advert {
-                        id: WireArtifact::PbId::proxy_encode(id),
-                        attribute: WireArtifact::PbAttribute::proxy_encode(attribute),
-                    })
+                    pb::slot_update::Update::Id(WireArtifact::PbId::proxy_encode(id))
                 }
             }),
         };
@@ -428,7 +422,6 @@ mod tests {
         async fn assemble_message<P: Peers + Send + 'static>(
             &self,
             _id: <U64Artifact as IdentifiableArtifact>::Id,
-            _attr: <U64Artifact as IdentifiableArtifact>::Attribute,
             _artifact: Option<(U64Artifact, NodeId)>,
             _peers: P,
         ) -> Result<(U64Artifact, NodeId), Aborted> {