+++ rs/pocket_ic_server/tests/test.rs
@@ -3,17 +3,26 @@ mod common;
 use crate::common::raw_canister_id_range_into;
 use candid::{Encode, Principal};
 use ic_agent::agent::{http_transport::ReqwestTransport, CallResponse};
+use ic_interfaces_registry::{
+    RegistryDataProvider, RegistryVersionedRecord, ZERO_REGISTRY_VERSION,
+};
 use ic_management_canister_types::ProvisionalCreateCanisterWithCyclesArgs;
 use ic_registry_proto_data_provider::ProtoRegistryDataProvider;
+use ic_registry_transport::pb::v1::{
+    registry_mutation::Type, RegistryAtomicMutateRequest, RegistryMutation,
+};
 use ic_utils::interfaces::ManagementCanister;
 use pocket_ic::common::rest::{
-    HttpGatewayBackend, HttpGatewayDetails, HttpsConfig, InstanceConfig, SubnetConfigSet,
+    CreateHttpGatewayResponse, HttpGatewayBackend, HttpGatewayConfig, HttpGatewayDetails,
+    HttpsConfig, InstanceConfig, SubnetConfigSet,
 };
 use pocket_ic::{PocketIc, PocketIcBuilder, WasmResult};
 use rcgen::{CertificateParams, KeyPair};
+use registry_canister::init::RegistryCanisterInitPayload;
 use reqwest::blocking::Client;
 use reqwest::Client as NonblockingClient;
 use reqwest::{StatusCode, Url};
+use slog::Level;
 use std::io::Read;
 use std::io::Write;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
@@ -27,6 +36,7 @@ pub const LOCALHOST: &str = "127.0.0.1";
 fn start_server_helper(
     parent_pid: Option<u32>,
     ttl: Option<u64>,
+    capture_stdout: bool,
     capture_stderr: bool,
 ) -> (Url, Child) {
     let bin_path = std::env::var_os("POCKET_IC_BIN").expect("Missing PocketIC binary");
@@ -49,6 +59,9 @@ fn start_server_helper(
     }
     if let Some(ttl) = ttl {
         cmd.arg("--ttl").arg(ttl.to_string());
+    }
+    if capture_stdout {
+        cmd.stdout(std::process::Stdio::piped());
     }
     if capture_stderr {
         cmd.stderr(std::process::Stdio::piped());
@@ -74,7 +87,7 @@ fn start_server_helper(
 
 pub fn start_server() -> Url {
     let parent_pid = std::os::unix::process::parent_id();
-    start_server_helper(Some(parent_pid), None, false).0
+    start_server_helper(Some(parent_pid), None, false, false).0
 }
 
 #[test]
@@ -98,6 +111,7 @@ fn test_creation_of_instance_extended() {
         .into(),
         state_dir: None,
         nonmainnet_features: false,
+        log_level: None,
     };
     let response = client
         .post(url.join("instances").unwrap())
@@ -450,7 +464,7 @@ fn test_specified_id() {
 
 #[test]
 fn test_dashboard() {
-    let (server_url, _) = start_server_helper(None, Some(5), true);
+    let (server_url, _) = start_server_helper(None, Some(5), false, false);
     let subnet_config_set = SubnetConfigSet {
         nns: true,
         application: 1,
@@ -510,14 +524,14 @@ const CANISTER_LOGS_WAT: &str = r#"
 "#;
 
 #[test]
-fn canister_logs() {
+fn canister_and_replica_logs() {
     const INIT_CYCLES: u128 = 2_000_000_000_000;
-    let (server_url, mut out) = start_server_helper(None, Some(5), true);
-    let subnet_config_set = SubnetConfigSet {
-        application: 1,
-        ..Default::default()
-    };
-    let pic = PocketIc::from_config_and_server_url(subnet_config_set, server_url);
+    let (server_url, mut out) = start_server_helper(None, Some(5), true, true);
+    let pic = PocketIcBuilder::new()
+        .with_application_subnet()
+        .with_server_url(server_url)
+        .with_log_level(Level::Info)
+        .build();
 
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, INIT_CYCLES);
@@ -525,6 +539,49 @@ fn canister_logs() {
     pic.install_canister(canister_id, canister_logs_wasm, vec![], None);
 
     drop(pic);
+
+    let mut stdout = String::new();
+    out.stdout
+        .take()
+        .unwrap()
+        .read_to_string(&mut stdout)
+        .unwrap();
+    assert!(stdout.contains("Finished executing install_code message on canister CanisterId(lxzze-o7777-77777-aaaaa-cai)"));
+
+    let mut stderr = String::new();
+    out.stderr
+        .take()
+        .unwrap()
+        .read_to_string(&mut stderr)
+        .unwrap();
+    assert!(stderr.contains("Logging works!"));
+}
+
+#[test]
+fn canister_and_no_replica_logs() {
+    const INIT_CYCLES: u128 = 2_000_000_000_000;
+    let (server_url, mut out) = start_server_helper(None, Some(5), true, true);
+    let pic = PocketIcBuilder::new()
+        .with_application_subnet()
+        .with_server_url(server_url)
+        .with_log_level(Level::Error)
+        .build();
+
+    let canister_id = pic.create_canister();
+    pic.add_cycles(canister_id, INIT_CYCLES);
+    let canister_logs_wasm = wat::parse_str(CANISTER_LOGS_WAT).unwrap();
+    pic.install_canister(canister_id, canister_logs_wasm, vec![], None);
+
+    drop(pic);
+
+    let mut stdout = String::new();
+    out.stdout
+        .take()
+        .unwrap()
+        .read_to_string(&mut stdout)
+        .unwrap();
+    assert!(!stdout.contains("Finished executing install_code message on canister CanisterId(lxzze-o7777-77777-aaaaa-cai)"));
+
     let mut stderr = String::new();
     out.stderr
         .take()
@@ -597,7 +654,7 @@ fn canister_state_dir() {
     let state_dir = TempDir::new().unwrap();
     let state_dir_path_buf = state_dir.path().to_path_buf();
 
-    // Create a PocketIC instance with NNS and app subets.
+    // Create a PocketIC instance with NNS and app subnets.
     let pic = PocketIcBuilder::new()
         .with_state_dir(state_dir_path_buf.clone())
         .with_nns_subnet()
@@ -665,7 +722,7 @@ fn canister_state_dir() {
     drop(pic);
 
     // Start a new PocketIC server.
-    let (new_server_url, _) = start_server_helper(None, Some(5), false);
+    let (new_server_url, _) = start_server_helper(None, Some(5), false, false);
 
     // Create a PocketIC instance mounting the state created so far.
     let pic = PocketIcBuilder::new()
@@ -694,7 +751,7 @@ fn canister_state_dir() {
     drop(pic);
 
     // Start a new PocketIC server.
-    let (newest_server_url, _) = start_server_helper(None, Some(5), false);
+    let (newest_server_url, _) = start_server_helper(None, Some(5), false, false);
 
     // Create a PocketIC instance mounting the NNS state created so far.
     let nns_subnet_seed = topology.0.get(&nns_subnet).unwrap().subnet_seed;
@@ -910,4 +967,243 @@ fn test_query_stats_live() {
             }
         }
     })
+}
+
+/// Tests subnet read state requests.
+#[test]
+fn test_subnet_read_state() {
+    const INIT_CYCLES: u128 = 2_000_000_000_000;
+
+    // Create PocketIC instance with one NNS subnet and one app subnet.
+    let mut pic = PocketIcBuilder::new()
+        .with_nns_subnet()
+        .with_application_subnet()
+        .build();
+
+    // Retrieve the app subnet from the topology.
+    let topology = pic.topology();
+    let app_subnet = topology.get_app_subnets()[0];
+
+    // We create a counter canister on the app subnet.
+    let canister_id = pic.create_canister_on_subnet(None, None, app_subnet);
+    pic.add_cycles(canister_id, INIT_CYCLES);
+    let counter_wasm = wat::parse_str(COUNTER_WAT).unwrap();
+    pic.install_canister(canister_id, counter_wasm, vec![], None);
+
+    let endpoint = pic.make_live(None);
+
+    let rt = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .unwrap();
+    rt.block_on(async {
+        let agent = ic_agent::Agent::builder()
+            .with_url(endpoint.clone())
+            .build()
+            .unwrap();
+        agent.fetch_root_key().await.unwrap();
+
+        let metrics = agent.read_state_subnet_metrics(app_subnet).await.unwrap();
+        assert_eq!(metrics.num_canisters, 1);
+    })
+}
+
+/// Tests that HTTP gateway can handle requests with IP address hosts.
+#[test]
+fn test_gateway_ip_addr_host() {
+    // Create PocketIC instance with one NNS subnet and one app subnet.
+    let mut pic = PocketIcBuilder::new()
+        .with_nns_subnet()
+        .with_application_subnet()
+        .build();
+
+    // Retrieve the app subnet from the topology.
+    let topology = pic.topology();
+    let app_subnet = topology.get_app_subnets()[0];
+
+    // We create a canister on the app subnet.
+    pic.create_canister_on_subnet(None, None, app_subnet);
+
+    let mut endpoint = pic.make_live(None);
+    endpoint
+        .set_ip_host(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)))
+        .unwrap();
+
+    let rt = tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .unwrap();
+    rt.block_on(async {
+        let agent = ic_agent::Agent::builder()
+            .with_url(endpoint.clone())
+            .build()
+            .unwrap();
+        agent.fetch_root_key().await.unwrap();
+
+        let metrics = agent.read_state_subnet_metrics(app_subnet).await.unwrap();
+        assert_eq!(metrics.num_canisters, 1);
+    })
+}
+
+#[test]
+fn test_unresponsive_gateway_backend() {
+    let client = Client::new();
+
+    // Create PocketIC instance with one NNS subnet and one app subnet.
+    let (backend_server_url, mut backend_process) = start_server_helper(None, None, false, false);
+    let pic = PocketIcBuilder::new()
+        .with_nns_subnet()
+        .with_application_subnet()
+        .with_server_url(backend_server_url.clone())
+        .build();
+
+    // Create HTTP gateway on a different gateway server.
+    let (gateway_server_url, _) = start_server_helper(None, None, false, false);
+    let create_gateway_endpoint = gateway_server_url.join("http_gateway").unwrap();
+    let backend_instance_url = backend_server_url
+        .join(&format!("instances/{}/", pic.instance_id()))
+        .unwrap();
+    let http_gateway_config = HttpGatewayConfig {
+        ip_addr: None,
+        port: None,
+        forward_to: HttpGatewayBackend::Replica(backend_instance_url.to_string()),
+        domains: None,
+        https_config: None,
+    };
+    let res = client
+        .post(create_gateway_endpoint)
+        .json(&http_gateway_config)
+        .send()
+        .unwrap()
+        .json::<CreateHttpGatewayResponse>()
+        .unwrap();
+    let endpoint = match res {
+        CreateHttpGatewayResponse::Created(info) => {
+            let port = info.port;
+            Url::parse(&format!("http://localhost:{}/", port)).unwrap()
+        }
+        CreateHttpGatewayResponse::Error { message } => {
+            panic!("Failed to crate http gateway: {}", message)
+        }
+    };
+
+    // Query the status endpoint via HTTP gateway.
+    let resp = client
+        .get(endpoint.join("api/v2/status").unwrap())
+        .send()
+        .unwrap();
+    assert_eq!(resp.status(), StatusCode::OK);
+
+    // Kill the backend server, but keep the HTTP gateway running.
+    drop(pic);
+    backend_process.kill().unwrap();
+
+    // Query the status endpoint via HTTP gateway again.
+    let resp = client
+        .get(endpoint.join("api/v2/status").unwrap())
+        .send()
+        .unwrap();
+    assert_eq!(resp.status(), StatusCode::BAD_GATEWAY);
+    assert!(String::from_utf8(resp.bytes().unwrap().as_ref().to_vec())
+        .unwrap()
+        .contains("connection_failure: client error"));
+}
+
+#[test]
+fn test_invalid_gateway_backend() {
+    // Create HTTP gateway with an invalid backend URL
+    let (gateway_server_url, _) = start_server_helper(None, None, false, false);
+    let create_gateway_endpoint = gateway_server_url.join("http_gateway").unwrap();
+    let backend_url = "http://240.0.0.0";
+    let http_gateway_config = HttpGatewayConfig {
+        ip_addr: None,
+        port: None,
+        forward_to: HttpGatewayBackend::Replica(backend_url.to_string()),
+        domains: None,
+        https_config: None,
+    };
+    let client = Client::new();
+    let res = client
+        .post(create_gateway_endpoint)
+        .json(&http_gateway_config)
+        .send()
+        .unwrap()
+        .json::<CreateHttpGatewayResponse>()
+        .unwrap();
+    match res {
+        CreateHttpGatewayResponse::Created(_info) => {
+            panic!("Suceeded to create http gateway!")
+        }
+        CreateHttpGatewayResponse::Error { message } => {
+            assert!(message.contains("An error happened during communication with the replica: error sending request for url"));
+        }
+    };
+}
+
+fn record_to_mutation(r: RegistryVersionedRecord<Vec<u8>>) -> RegistryMutation {
+    let mut m = RegistryMutation::default();
+
+    let t = if r.value.is_none() {
+        Type::Delete
+    } else {
+        Type::Insert
+    };
+    m.set_mutation_type(t);
+    m.key = r.key.as_bytes().to_vec();
+    m.value = r.value.unwrap_or_default();
+
+    m
+}
+
+#[test]
+fn registry_canister() {
+    // Create a temporary state directory persisted throughout the test.
+    let state_dir = TempDir::new().unwrap();
+    let state_dir_path_buf = state_dir.path().to_path_buf();
+
+    // Create a PocketIC instance with NNS, II and two app subnets.
+    let pic = PocketIcBuilder::new()
+        .with_state_dir(state_dir_path_buf.clone())
+        .with_nns_subnet()
+        .with_ii_subnet()
+        .with_application_subnet()
+        .with_application_subnet()
+        .build();
+
+    // Encode the local registry into a registry canister initial payload.
+    let registry_proto_path = state_dir_path_buf.join("registry.proto");
+    let registry_data_provider = ProtoRegistryDataProvider::load_from_file(registry_proto_path);
+    let updates = registry_data_provider
+        .get_updates_since(ZERO_REGISTRY_VERSION)
+        .unwrap();
+    let mutations = updates
+        .into_iter()
+        .map(record_to_mutation)
+        .collect::<Vec<RegistryMutation>>();
+    let mutate_request = RegistryAtomicMutateRequest {
+        mutations,
+        ..Default::default()
+    };
+    let registry_init_payload = RegistryCanisterInitPayload {
+        mutations: vec![mutate_request],
+    };
+
+    // Create the registry canister.
+    let registry_canister_id = Principal::from_text("rwlgt-iiaaa-aaaaa-aaaaa-cai").unwrap();
+    let actual_registry_canister_id = pic
+        .create_canister_with_id(None, None, registry_canister_id)
+        .unwrap();
+    assert_eq!(registry_canister_id, actual_registry_canister_id);
+
+    // Install the registry canister.
+    let registry_path = std::env::var_os("REGISTRY_WASM")
+        .expect("Missing REGISTRY_WASM (path to REGISTRY wasm) in env.");
+    let registry_canister_wasm =
+        std::fs::read(registry_path).expect("Could not read REGISTRY wasm file.");
+    pic.install_canister(
+        registry_canister_id,
+        registry_canister_wasm,
+        Encode!(&registry_init_payload).unwrap(),
+        None,
+    );
 }