+++ rs/sns/integration_tests/src/lib.rs
@@ -51,3 +51,6 @@ mod manage_ledger_parameters;
 
 #[cfg(test)]
 mod manage_dapp_canister_settings;
+
+#[cfg(test)]
+mod http_request;