+++ rs/state_machine_tests/src/tests.rs
@@ -1,4 +1,4 @@
-use ic_crypto_ecdsa_secp256k1::{DerivationIndex, DerivationPath, PrivateKey, PublicKey};
+use ic_crypto_secp256k1::{DerivationIndex, DerivationPath, PrivateKey, PublicKey};
 use proptest::{collection::vec as pvec, prelude::*, prop_assert, proptest};
 
 proptest! {