+++ rs/types/types/src/consensus.rs
@@ -19,9 +19,9 @@ use ic_protobuf::{
     proxy::{try_from_option_field, ProxyDecodeError},
 };
 use serde::{Deserialize, Serialize};
+use std::cmp::PartialOrd;
 use std::convert::TryInto;
 use std::hash::Hash;
-use std::{cmp::PartialOrd, convert::Infallible};
 
 pub mod block_maker;
 pub mod catchup;
@@ -906,11 +906,9 @@ pub enum ConsensusMessage {
 impl IdentifiableArtifact for ConsensusMessage {
     const NAME: &'static str = "consensus";
     type Id = ConsensusMessageId;
-    type Attribute = ();
     fn id(&self) -> Self::Id {
         self.get_id()
     }
-    fn attribute(&self) -> Self::Attribute {}
 }
 
 impl PbArtifact for ConsensusMessage {
@@ -918,8 +916,6 @@ impl PbArtifact for ConsensusMessage {
     type PbIdError = ProxyDecodeError;
     type PbMessage = ic_protobuf::types::v1::ConsensusMessage;
     type PbMessageError = ProxyDecodeError;
-    type PbAttribute = ();
-    type PbAttributeError = Infallible;
 }
 
 impl From<ConsensusMessage> for pb::ConsensusMessage {